bigdata_btl/requirements.txt
requests 
kafka-python
pyspark
elasticsearch
fastapi
uvicorn
torch